nix 0117

+++ src/main/java/com/douane/managed/bean/GACBean.java
@@ -383,7 +383,7 @@ public class GACBean {
 
         ListIterator<Materiel> it = this.getListMaterielByDet().listIterator();
         if (it!=null) {
-            this.setTotal((Float)0);
+            this.setTotal(0.0);
            while(it.hasNext()){
              setTotal(this.total+(Float)(it.next().getPu()));
           }
